@@ -1,15 +1,15 @@
 {
   "dependencies": {
-    "@angular/animations": "^19.2.8",
-    "@angular/cdk": "^19.2.11",
-    "@angular/common": "^19.2.8",
-    "@angular/compiler": "^19.2.8",
-    "@angular/core": "^19.2.8",
-    "@angular/forms": "^19.2.8",
-    "@angular/material": "^19.2.11",
-    "@angular/platform-browser": "^19.2.8",
-    "@angular/platform-browser-dynamic": "^19.2.8",
-    "@angular/router": "^19.2.8",
+    "@angular/animations": "^20.3.1",
+    "@angular/cdk": "^20.2.4",
+    "@angular/common": "^20.3.1",
+    "@angular/compiler": "^20.3.1",
+    "@angular/core": "^20.3.1",
+    "@angular/forms": "^20.3.1",
+    "@angular/material": "^20.2.4",
+    "@angular/platform-browser": "^20.3.1",
+    "@angular/platform-browser-dynamic": "^20.3.1",
+    "@angular/router": "^20.3.1",
     "@awesome-cordova-plugins/core": "^6.16.0",
     "@awesome-cordova-plugins/file": "^6.16.0",
     "@capacitor-community/keep-awake": "^7.0.0",
@@ -88,15 +88,15 @@
     "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.9",
-    "@angular-eslint/builder": "^19.3.0",
-    "@angular-eslint/eslint-plugin": "^19.3.0",
-    "@angular-eslint/eslint-plugin-template": "^19.3.0",
-    "@angular-eslint/schematics": "^19.3.0",
-    "@angular-eslint/template-parser": "^19.3.0",
-    "@angular/cli": "^19.2.9",
-    "@angular/compiler-cli": "^19.2.8",
-    "@angular/language-service": "^19.2.8",
+    "@angular-eslint/builder": "^20.3.0",
+    "@angular-eslint/eslint-plugin": "^20.3.0",
+    "@angular-eslint/eslint-plugin-template": "^20.3.0",
+    "@angular-eslint/schematics": "^20.3.0",
+    "@angular-eslint/template-parser": "^20.3.0",
+    "@angular/build": "^20.3.2",
+    "@angular/cli": "^20.3.2",
+    "@angular/compiler-cli": "^20.3.1",
+    "@angular/language-service": "^20.3.1",
     "@capacitor/cli": "^7.2.0",
     "@tailwindcss/postcss": "^4.1.4",
     "@trapezedev/project": "^7.1.3",
@@ -111,9 +111,9 @@
     "@types/platform": "^1.3.6",
     "@types/proj4": "^2.5.5",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.31.0",
-    "@typescript-eslint/parser": "^8.31.0",
-    "eslint": "^9.25.1",
+    "@typescript-eslint/eslint-plugin": "^8.33.1",
+    "@typescript-eslint/parser": "^8.33.1",
+    "eslint": "^9.28.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsdoc": "^50.6.11",
     "eslint-plugin-prefer-arrow": "^1.2.3",
@@ -128,6 +128,7 @@
     "karma-jasmine-html-reporter": "^2.1.0",
     "karma-junit-reporter": "^2.0.1",
     "patch-package": "^8.0.0",
+    "postcss": "^8.5.3",
     "puppeteer": "^24.7.2",
     "replace-in-file": "^8.3.0",
     "tailwindcss": "^4.1.4",
