@@ -2,7 +2,7 @@ import { Component, inject } from "@angular/core";
 import { Dir } from "@angular/cdk/bidi";
 import { MatButton, MatAnchor } from "@angular/material/button";
 import { CdkScrollable } from "@angular/cdk/scrolling";
-import { NgIf, NgFor } from "@angular/common";
+
 import { MatHint } from "@angular/material/form-field";
 import { MatDialog, MatDialogRef, MatDialogTitle, MatDialogClose, MatDialogContent } from "@angular/material/dialog";
 import { Angulartics2OnModule } from "angulartics2";
@@ -21,7 +21,7 @@ import type { DataContainer } from "../../models";
 @Component({
     selector: "files-share-dialog",
     templateUrl: "./files-shares-dialog.component.html",
-    imports: [Dir, MatDialogTitle, MatButton, MatDialogClose, CdkScrollable, MatDialogContent, Angulartics2OnModule, MatAnchor, NgIf, MatHint, NgFor]
+    imports: [Dir, MatDialogTitle, MatButton, MatDialogClose, CdkScrollable, MatDialogContent, Angulartics2OnModule, MatAnchor, MatHint]
 })
 export class FilesSharesDialogComponent {
 
