@@ -1,6 +1,6 @@
 import { Component, ElementRef, AfterViewInit, HostListener, inject, viewChild } from "@angular/core";
 import { Dir } from "@angular/cdk/bidi";
-import { NgIf, NgFor, NgStyle, NgClass } from "@angular/common";
+import { NgStyle, NgClass } from "@angular/common";
 import { MatButton, MatAnchor, MatIconButton } from "@angular/material/button";
 import { CdkScrollable } from "@angular/cdk/scrolling";
 import { MatFormField, MatLabel, MatSuffix } from "@angular/material/form-field";
@@ -42,7 +42,7 @@ interface PrivatePoiEditDialogData {
 @Component({
     selector: "private-poi-edit-dialog",
     templateUrl: "private-poi-edit-dialog.component.html",
-    imports: [Dir, MatDialogTitle, NgIf, MatButton, MatDialogClose, CdkScrollable, MatDialogContent, MatFormField, MatLabel, MatInput, FormsModule, MatAnchor, ImageCaptureDirective, Angulartics2OnModule, MatTooltip, NgFor, NgStyle, NgClass, MatIconButton, MatSuffix, CoordinatesComponent, MatDialogActions, MatMenu, MatMenuItem, MatMenuTrigger]
+    imports: [Dir, MatDialogTitle, MatButton, MatDialogClose, CdkScrollable, MatDialogContent, MatFormField, MatLabel, MatInput, FormsModule, MatAnchor, ImageCaptureDirective, Angulartics2OnModule, MatTooltip, NgStyle, NgClass, MatIconButton, MatSuffix, CoordinatesComponent, MatDialogActions, MatMenu, MatMenuItem, MatMenuTrigger]
 })
 export class PrivatePoiEditDialogComponent implements AfterViewInit {
     private static readonly NUMBER_OF_ICONS_PER_ROW = 4;
